# Fix performance degradation due to metrics