### Description
The gcs/read_bytes_count metric is single-handedly responsible for perf degradation due when metrics are enabled. With this change, we are now using an atomic Int64 which efficiently computes the metric. The Int64Counter that uses OTel for metric computation has now been replaced with Int64ObservableCounter that uses the aforementioned atomic Int64 to report the same metric.

Perf:
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
|--------|------------|--------------|------------|--------------|--------------|
| Master |  0.25MiB   | 528.52MiB/s  | 1.32MiB/s  |  74.95MiB/s  |  1.15MiB/s   |
|   PR   |  0.25MiB   | 547.94MiB/s  | 1.34MiB/s  |  78.84MiB/s  |  1.31MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 3587.54MiB/s | 76.99MiB/s | 1485.06MiB/s |  84.21MiB/s  |
|   PR   | 48.828MiB  | 3659.14MiB/s | 86.45MiB/s | 1483.11MiB/s |  84.15MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 3635.5MiB/s  | 77.45MiB/s | 760.01MiB/s  |  29.29MiB/s  |
|   PR   | 976.562MiB | 3559.9MiB/s  | 78.1MiB/s  | 921.71MiB/s  |  28.44MiB/s  |

### Link to the issue in case of a bug fix.
b/386226641

### Testing details
1. Manual - Verified that the metric is being correctly computed. Also verified that the performance is the same with and without metrics being enabled.
2. Unit tests - NA - the functionality of the module is not changing - only the internal implementation changes.
3. Integration tests - NA - the functionality is not changing - only the internal implementation changes.